% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_pdf.R
\name{py_count_tokens}
\alias{py_count_tokens}
\title{Count Tokens}
\usage{
py_count_tokens(x)
}
\description{
Use python tiktoken package to count the tokens in full text for a given model
}
\examples{
py_pdf_blocks("data-raw/onedrive/2024 35 2/00-Prospects of Deterrence_ Deterrence Theory_ Representation and Evidence.pdf") |>
  process_blocks() |>
  purrr::pluck("full_text") |>
  py_count_tokens()
}
