% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_pdf.R
\name{py_pdf_blocks}
\alias{py_pdf_blocks}
\title{Extract pdf blocks}
\usage{
py_pdf_blocks(filepath)
}
\arguments{
\item{filepath}{path to pdf file}
}
\value{
pdf text
}
\description{
Use pymupdf to extract text blocks from pdf document
}
\examples{
\dontrun{
  py_pdf_blocks("paper.pdf")
}
}
