% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_python_env}
\alias{setup_python_env}
\title{Setup Python Environment}
\usage{
setup_python_env(method = "virtualenv", envname = "r-myenv")
}
\arguments{
\item{method}{Either "virtualenv" or "conda"}

\item{envname}{Name of the environment}
}
\description{
Creates and configures a virtualenv or conda env and installs required Python packages.
}
