% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpe-package.R
\docType{package}
\name{dpe-package}
\alias{dpe}
\alias{dpe-package}
\title{Defence and Peace Economics}
\description{
Automated evidence synthesis from the journal of defence and peace economics.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
