% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_pdf.R
\name{py_pdf_blocks}
\alias{py_pdf_blocks}
\title{Extract pdf blocks}
\usage{
py_pdf_blocks(filepath)
}
\description{
Use pymupdf to extract text blocks from pdf document
}
\examples{
py_pdf_blocks("data-raw/onedrive/2024 35 2/00-Prospects of Deterrence_ Deterrence Theory_ Representation and Evidence.pdf")
}
